requirements.txt
pandas==1.5.3
python-docx==0.8.11
openpyxl==3.1.2
pdf2image==1.16.3
pytesseract==0.3.10
pillow==9.5.0
opencv-python-headless==4.8.0.74
pdfplumber==0.9.0
streamlit==1.24.0
